==> BinaryClassifierMethod/1_CreatePositiveAID.ipynb <==
Fetches those customers who bought atleast 1 product in target week

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

### Objective

To train the model, we need the data in the following format:

Columns: [features], relevance

Column Descriptions:
2. features: features of user, product etc
3. relevance of the product (1 = bought, 0 not bought)

Breakdown to 1 user:

We need to create examples of feature and relevance essentially. This basically means each row is a pairwise object.
We need relevance to have positive samples and negative samples.

Positive samples: Items the user bought

Negative samples: Items the user didn't buy

How to create the data then?

1. Only create data for those users who bought something: i.e. get customers who bought something
2. For each of those customers, create a list of the 12 products they bought, these are the positive samples
3. For each of those customers, create a list of products they didn't buy, these are the negative samples

Repeat above for all users.

### Creating the samples

df = pd.read_parquet('../data/train.parquet')

# predicting what they bought next week (so train on data before this)
d_end = datetime.datetime(2019, 9, 29).date()
d_start = datetime.datetime(2019, 9, 23).date()

tdf = df[(df['date'] >= d_start) & (df['date'] <= d_end)].reset_index(drop=True).copy()

tdf['a_count'] = tdf.groupby(['cust_id', 'article_id'])['article_id'].transform('count')

tdf.sort_values(['cust_id', 'date', 'a_count'], ascending=[True, False, False], inplace=True)

bdf = tdf.groupby(["cust_id"])["article_id"].agg(
    lambda x: list(x.values[np.sort(np.unique(x.values, return_index=True)[1])])).reset_index()

bdf

bdf.to_parquet('../data/bought_articles_in_order.parquet')
