==> BinaryClassifierMethod/2a_CreateUserEmbedding.ipynb <==
Creates user embeddings

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

df = pd.read_parquet('../data/train.parquet')

with open('../data/emb_map.pkl', 'rb') as handle:
    emb_map = pkl.load(handle)

d_end = datetime.datetime(2019, 9, 22).date()

trdf = df[(df['date'] <= d_end)].reset_index(drop=True).copy()

trdf.shape

bdf = pd.read_parquet('../data/bought_articles_in_order.parquet')

cbs = set(bdf['cust_id'])

trdf = trdf[trdf['cust_id'].isin(cbs)].reset_index(drop=True).copy()

trdf.shape

trdf['cust_id'].nunique()

len(cbs)

trdf.sort_values(['cust_id', 'date'], ascending=[True, True], inplace=True)

cids = trdf['cust_id'].unique()
cgroups = trdf[['article_id', 'cust_id']].groupby('cust_id')

cemb_map = {}

def moveVector(sourceVector, destinationVector, direction, magnitude = None):
    magnitude = magnitude if magnitude is not None else 0.5
    new_q = sourceVector + direction*magnitude*(destinationVector - sourceVector)
    return new_q

#x = A
#x = A + 0.5(B-A) = 0.5A + 0.5B
#x = 0.5A + 0.5B + 0.5(C - (0.5A + 0.5B)) = 0.5A +0.5B + 0.5(C-0.5A-0.5B) = 0.5A + 0.5B + 0.5C -.25A -0.25B = 0.25A + 0.25B + 0.5C

def getCustEmb(purchases):
    statevector = None
    for i, v in enumerate(purchases):
        if i == 0:
            statevector = v
        else:
            statevector = moveVector(statevector, v, direction=1, magnitude=0.5)
    
    statevector = statevector / np.linalg.norm(statevector)
    return statevector

all_embs = []
for cid in tqdm(cids, total=len(cids)):
    curr_group = cgroups.get_group(cid)
    keys = curr_group['article_id'].values
    embs = [emb_map[x] for x in keys]
    embs = np.array(embs)
    cemb = getCustEmb(embs)
    all_embs.append(cemb)
    cemb_map[cid] = cemb

len(all_embs)

with open('../data/cemb_map.pkl', 'wb') as handle:
    pkl.dump(cemb_map, handle, protocol=3)
