==> BinaryClassifierMethod/5a_Modelling.ipynb <==
Trains a 2 head neural network classifier to predict whether a user will purchase a paired product

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
from tqdm.notebook import tqdm
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.preprocessing import StandardScaler
from tensorflow import keras
from tensorflow.keras import layers
import tensorflow as tf
import tensorflow_addons as tfa
from tensorflow.keras import regularizers
import pickle as pkl

df = pd.read_parquet('../data/sampleTrain.parquet')

sss = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=0)

train_indices = []
val_indices = []
for train_index, val_index in sss.split(df, df['Y']):
    train_indices.append(train_index)
    val_indices.append(val_index)

train_df = df.iloc[train_indices[0]].reset_index(drop=True).copy()
val_df = df.iloc[val_indices[0]].reset_index(drop=True).copy()

train_df.head()

x_train = train_df.drop(['Y','Q'], axis=1).copy()
x_val = val_df.drop(['Y','Q'], axis=1).copy()

y_train = train_df['Y'].values
y_val = val_df['Y'].values

cf_xtrain = x_train[['c_' + str(x) for x in range(512)]].copy()
cf_xval = x_val[['c_' + str(x) for x in range(512)]].copy()

af_xtrain = x_train[['a_' + str(x) for x in range(512)]].copy()
af_xval = x_val[['a_' + str(x) for x in range(512)]].copy()

BATCH_SIZE = 64

fold_train_loss = []
fold_val_loss = []

train_dataset = tf.data.Dataset.from_tensor_slices(((cf_xtrain, af_xtrain), y_train))
val_dataset = tf.data.Dataset.from_tensor_slices(((cf_xval, af_xval), y_val))

train_dataset = train_dataset.batch(BATCH_SIZE)
val_dataset = val_dataset.batch(BATCH_SIZE)

train_dataset.prefetch(tf.data.AUTOTUNE)
val_dataset.prefetch(tf.data.AUTOTUNE)

def get_model():
    cf_inputs = tf.keras.Input((512, ))
    af_inputs = tf.keras.Input((512, ))
    
    cf_x = layers.Dense(25, activation='relu')(cf_inputs)
    cf_x = layers.Dense(12, activation='relu')(cf_x)
    
    af_x = layers.Dense(25, activation='relu')(af_inputs)
    af_x = layers.Dense(12, activation='relu')(af_x)
    
    fx = layers.Concatenate(axis=1)([cf_x, af_x])
    fx = layers.Dense(20, activation='relu', kernel_regularizer=regularizers.l2(0.001))(fx)
    fx = layers.Dense(10, activation='relu', kernel_regularizer=regularizers.l2(0.001))(fx)
    out = layers.Dense(1, activation='sigmoid', kernel_regularizer=regularizers.l2(0.001))(fx)
    
    model = tf.keras.Model(inputs=[cf_inputs, af_inputs], outputs=out)
    model.compile(loss='binary_crossentropy',
                optimizer=tf.keras.optimizers.SGD(0.01),
                 metrics=['accuracy'])
    return model

keras.backend.clear_session()
dnn = get_model()
dnn.summary()

total = len(y_train)
pos = sum(y_train)
neg = total-pos

weight_for_0 = (1 / neg) * (total / 2.0)
weight_for_1 = (1 / pos) * (total / 2.0)

class_weight = {0: weight_for_0, 1: weight_for_1}

print('Weight for class 0: {:.2f}'.format(weight_for_0))
print('Weight for class 1: {:.2f}'.format(weight_for_1))

1 - sum(y_train)/len(y_train)

1 - sum(y_val)/len(y_val)

history = dnn.fit(
    train_dataset,
    validation_data=val_dataset,
    verbose=1, epochs=30, workers=1, shuffle=False, class_weight=class_weight)


dnn.save('../models/dnn1')

preds = dnn.predict(val_dataset, verbose=1)

preds = preds.flatten()

preds[0]

y_val[0]

preds[:40]
