==> BinaryClassifierMethod/6b_CreateFinalPredictions.ipynb <==
Creates the final prediction file to submit to kaggle

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter
import pickle as pkl

with open('../data/P.pkl', 'rb') as handle:
    P = pkl.load(handle)

with open('../data/A.pkl', 'rb') as handle:
    A = pkl.load(handle)

with open('../data/C.pkl', 'rb') as handle:
    C = pkl.load(handle)

pdf = pd.DataFrame()
pdf['score'] = P 
pdf['article_id'] = A
pdf['cust_id'] = C

pdf.head()

pdf.head()

pdf.shape

pdf.sort_values(['cust_id', 'score'], ascending=[True, False], inplace=True)

pdf.head()

gdf = pdf.groupby('cust_id')['article_id'].agg(lambda x: list(x.values[:12]))

pdf.head()

gdf = gdf.reset_index().copy()

def createPredFormat(x):
    preds_sep_f = ''
    for i in range(len(x)):
        val = '0' + str(x[i])
        if i != len(x) - 1:
            val += ' '
        preds_sep_f += val
    return preds_sep_f

gdf['prediction'] = gdf['article_id'].apply(lambda x: createPredFormat(x))

gdf.drop('article_id', axis=1, inplace=True)

gdf

mfa = [751471001, 909370001, 918522001, 924243001, 918292001, 915526001,
            448509014, 915529003, 751471043, 706016001, 865799006, 863595006,
            850917001, 762846027, 898694001, 673677002, 714790020, 896152002,
            863583001, 762846031, 919365008, 852584001, 923758001, 715624001]

mfa = createPredFormat(mfa)

with open('../data/cid_map.pkl', 'rb') as handle:
    cid_map = pkl.load(handle)

len(cid_map)

sdf = pd.read_csv('../data/sample_submission.csv')

sdf['cust_id'] = sdf['customer_id'].apply(lambda x: cid_map[x])

sdf.drop('prediction', axis=1, inplace=True)

sdf = sdf.merge(right=gdf, on='cust_id', how='left')

sdf.isna().sum()

sdf['prediction'].fillna(mfa, inplace=True)

sdf.isna().sum()

sdf.drop('cust_id', axis=1, inplace=True)

sdf

sdf.to_csv('sub_nn_binary.csv', index=False)
